fix: deleted ./pages/_types.ts
Next will throw error if ./pages folder contains any file without a
React Component exported when build. The types have been moved into the
index page file.
